feat: refactor backend configuration tests and introduce helper functions for improved mock setup
